requirements.txt
# Core Dependencies
flask==2.3.3
python-dotenv==0.21.1
pyyaml==6.0.2
requests==2.31.0
python-jose[cryptography]==3.3.0
pycryptodome==3.19.0
cryptography==42.0.0
bcrypt==4.0.1
pyopenssl==24.0.0

# Security and Encryption
cryptography==42.0.0
bcrypt==4.0.1
pycryptodome==3.19.0
pyopenssl==24.0.0

# Database
pymongo==4.10.1
SQLAlchemy==2.0.23
alembic==1.12.1

# Real-time Communication
eventlet==0.33.3
python-socketio==5.10.0
python-engineio==4.9.0
aiohttp==3.10.11
aiohttp-cors>=0.7.0
websockets==12.0

# Network Tools
scapy==2.6.1
netifaces
psutil==5.9.6
dnspython==2.4.2
pysocks==1.7.1

# Scheduling and Async
apscheduler==3.10.4
trio==0.22.0
pyzmq==26.4.0 --only-binary :all:
pyasn1==0.6.1

# Protocols
paho-mqtt==2.1.0
aiocoap==0.4.7

# Monitoring and Logging
psutil==5.9.6
pyasn1-modules==0.4.1

# Error handling and validation
python-dateutil==2.8.2
marshmallow==3.21.0

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# Monitoring Tools
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
grafana-api==1.0.1
influxdb-client==1.37.0
elasticsearch==8.11.1
kafka-python==2.0.2

# Performance Monitoring
memory-profiler==0.61.0
aiomonitor>=0.6.0

# Security Monitoring
python-nmap

# Testing
tox==4.12.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
beautifulsoup4==4.12.2
flask-jwt-extended==4.3.1
flask-cors==4.0.0
shodan

# OSINT Packages

ratelimit
tenacity==8.2.3
python-jose==3.3.0
flask-login==0.6.3
werkzeug==2.3.7
itsdangerous==2.1.2
jinja2>=3.1.2
phonenumbers==8.13.13

# Async IO packages
aiosmtplib==2.0.1
aiodns==3.0.0
aiomysql==0.1.1
async-timeout==4.0.2
aioredis==2.0.1
asyncssh==2.12.0
aiosignal==1.3.1
asyncio-mqtt==0.16.1
asyncio-redis==0.16.0
aiosqlite==0.19.0

# GSM/SS7 packages

gsm0338==1.0.0
pyserial==3.5
aioudp==0.1.0
pyModbusTCP==0.2.1
paramiko==2.12.0
pyotp==2.9.0

pyrdp==0.1.6
pysnmp==4.4.12
aiosmtpd==1.4.6
geoip2==4.7.0
maxminddb==2.6.3
